+++ Tools/SandvikProbe/obj/Debug/net10.0/SandvikProbe.AssemblyInfo.cs
@@ -13,7 +13,7 @@ using System.Reflection;
 [assembly: System.Reflection.AssemblyCompanyAttribute("SandvikProbe")]
 [assembly: System.Reflection.AssemblyConfigurationAttribute("Debug")]
 [assembly: System.Reflection.AssemblyFileVersionAttribute("1.0.0.0")]
-[assembly: System.Reflection.AssemblyInformationalVersionAttribute("1.0.0+7b90b4ea6f4d16073d6853544e8638dca31af656")]
+[assembly: System.Reflection.AssemblyInformationalVersionAttribute("1.0.0+e953325c73d3504f7a544005fb7bf4cc53880bf0")]
 [assembly: System.Reflection.AssemblyProductAttribute("SandvikProbe")]
 [assembly: System.Reflection.AssemblyTitleAttribute("SandvikProbe")]
 [assembly: System.Reflection.AssemblyVersionAttribute("1.0.0.0")]